+++ resources/agent_ruby.rb
@@ -7,8 +7,8 @@
 actions :install, :remove
 default_action :install
 
-attribute :license, :kind_of => String, :default => nil
-attribute :agent_type, :kind_of => String, :required => true, :default => 'ruby'
+attribute :license, :kind_of => String, :default => NewRelic.application_monitoring_license(node)
+attribute :agent_type, :kind_of => String, :default => 'ruby'
 attribute :install_dir, :kind_of => String, :default => '/opt/newrelic/ruby'
 attribute :app_user, :kind_of => String, :default => 'newrelic'
 attribute :app_group, :kind_of => String, :default => 'newrelic'
@@ -36,6 +36,7 @@ attribute :daemon_proxy_user, :kind_of => String, :default => nil
 attribute :daemon_proxy_password, :kind_of => String, :default => nil
 attribute :capture_params, :kind_of => String, :default => nil
 attribute :ignored_params, :kind_of => String, :default => nil
+attribute :enable_custom_tracing, :kind_of => [TrueClass, FalseClass], :default => false
 attribute :transaction_tracer_enable, :kind_of => [TrueClass, FalseClass], :default => true
 attribute :transaction_tracer_threshold, :kind_of => String, :default => nil
 attribute :transaction_tracer_record_sql, :kind_of => String, :default => nil